# speed up `permute_once()`